+++ src/types/cwe.ts
@@ -28,15 +28,11 @@ import { isPositiveInteger } from './integer'
  * value `399` might represent `https://cwe.mitre.org/data/definitions/399.html`.
  *
  * @see {@link isCWE}
- *
- * @beta
  */
 export type CWE = PositiveInteger
 
-/** @beta */
 export function isCWE (value: any): value is CWE {
   return isPositiveInteger(value)
 }
 
-/** @beta */
 export class CweRepository extends SortableNumbers<CWE> {}